добавил аутентификацию

+++ register.php
@@ -1,4 +1,5 @@
-<?php session_start();?>
+<?php session_start();
+echo json_encode($_SESSION)?>
 <!DOCTYPE html>
 <html lang="ru">
 <head>
@@ -16,16 +17,21 @@
         </div>
     </header>
     <main>
+      <!--https://github.com/RayStell  -->
     <?php 
-         function showError($field){
-            $listErrors = $_SESSION['register-errors'];
-
+         function showError($field) {
+            if (!array_key_exists('register-errors', $_SESSION)){
+                echo '';
+            } else {
+                $listErrors = $_SESSION['register-errors'];
+                
                 if (array_key_exists($field, $listErrors)){
-                    $error = implode (',', $listErrors[$field]); 
-                    
-                     echo "<span class='error'>$error </span>";
+                    $error = implode (',', $listErrors[$field]);
+
+                    echo "<span class='error'> $error </span>";
                 }
-            } 
+            }
+         }
         ?>
         <section>
             <form method="POST" action="api/registrationUser.php" method="POST">
@@ -34,7 +40,6 @@
                 <label for="email"> Email <?php showError('email'); ?></label>
                 <input name="email" type="email" name="email" id="email" placeholder="example">
                 
-
                 <label for="name">Name <?php showError('name'); ?></label>
                 <input name="name" type="name" name="name" id="name" placeholder="example">
                 
@@ -50,6 +55,8 @@
                 <label for="password-confirm">repeat password <?php showError('password-confirm'); ?></label>
                 <input name="password-confirm" type="password"  id="password" placeholder="password">         
                
+                <label for="checkbox">Согласен на обработку персоналки <?php showError('agree'); ?></label>
+                <input name="agree" type="checkbox"  id="checkbox">         
                 <button type="submit">Регистрация</button>
             </form>
         </section>
